requirements.txt
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.7.1
idna==3.3
jieba3k==0.35.1
joblib==1.1.0
lxml==4.9.0
newspaper3k==0.2.8
nltk==3.7
Pillow==9.1.1
PyMuPDF==1.19.6
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.6.2
requests==2.28.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
tinysegmenter==0.3
tldextract==3.3.0
tqdm==4.64.0
urllib3==1.26.9
